Corrected Errors in Verification, added logging when it fails
The errors were logical once. For vesting the remaining amount
must be calculated from the the points where we fetched them. This
was not the case, used the latest block instead.

Added extensice logging when something fails, in order to reconstruct
occurenceCorrected Errors in Verification, added logging when it fails

The errors were logical once. For vesting the remaining amount
must be calculated from the the points where we fetched them. This
was not the case, used the latest block instead.

Added extensice logging when something fails, in order to reconstruct
occurence.