Adds logging to file, adds profile and homedir
This commit changes how the CLI starts. At statup the cli will check
if a profile is available. (currently only indicating the base dir)
If yes, use it, if not, create default with base dir = home.

in the base dir there will be a folder created (if not present) and
this folder will contain logs, and command specific outputs.