Possibility to append data from a csv file
This change enables the crowdloan command, to extend the contributions data
with contributions froma csv file.
If the additional contributors are already part of the contributions
the additional amount will be added to their existing one.
If they are not already part of this, they will be added with
the given amount.